+++ src/go/main.go
@@ -1,7 +1,7 @@
 package main
 
-import "wasm-components-with-wasmtime-embedding/go/wat"
+import "wasm-components-with-wasmtime-embedding/go/wasm"
 
 func main() {
-	wat.ProcessWat()
+	wasm.ProcessWasm()
 }
